Integrate PhotoListPage
